+++ src/Injector.js
@@ -3,11 +3,39 @@ import generateFingerprint from './Fingerprint'
 import { joinURI, normalizeURI } from './URI'
 import { retrieveIcons, parseIcons } from './Icons'
 
+const voidTags = [
+  'area',
+  'base',
+  'br',
+  'col',
+  'embed',
+  'hr',
+  'img',
+  'input',
+  'keygen',
+  'link',
+  'menuitem',
+  'meta',
+  'param',
+  'source',
+  'track',
+  'wbr'
+]
+
+const appleTags = {
+  'apple-touch-icon': 'link',
+  'apple-touch-startup-image': 'link',
+  'apple-mobile-web-app-title': 'meta',
+  'apple-mobile-web-app-capable': 'meta',
+  'apple-mobile-web-app-status-bar-style': 'meta'
+}
+
 function manifest (options, publicPath, icons, callback) {
   const content = Object.assign({ icons }, options)
   delete content.filename
   delete content.inject
   delete content.fingerprints
+  delete content.ios
   const json = JSON.stringify(content, null, 2)
   const filename = path.parse(options.filename)
   const output = options.fingerprints ? `${filename.name}.${generateFingerprint(json)}${filename.ext}` : `${filename.name}${filename.ext}`
@@ -46,4 +74,118 @@ export function injectResources (compilation, assets, callback) {
     }
   }
   callback()
+}
+
+export function generateAppleTags (options, assets) {
+  let tags = {}
+  if (options.ios) {
+    let apple = Object.assign({
+      'apple-mobile-web-app-title': options.name,
+      'apple-mobile-web-app-capable': 'yes',
+      'apple-mobile-web-app-status-bar-style': 'default'
+    }, typeof options.ios === 'object' ? options.ios : {})
+    for (let tag in apple) {
+      const type = appleTags[tag]
+      if (!type) continue // not a valid apple tag
+      applyTag(tags, type, formatAppleTag(tag, apple[tag]))
+    }
+    if (assets) {
+      for (let asset of assets) {
+        if (asset.ios && asset.ios.valid) {
+          if (asset.ios.valid === 'default') {
+            applyTag(tags, 'link', {
+              rel: 'apple-touch-startup-image',
+              href: asset.ios.href
+            })
+          } else {
+            applyTag(tags, 'link', {
+              rel: 'apple-touch-icon',
+              sizes: asset.ios.size,
+              href: asset.ios.href
+            })
+          }
+        }
+      }
+    }
+  }
+  return tags
+}
+
+function formatAppleTag (tag, content) {
+  if (tag === 'apple-touch-icon') {
+    if (typeof content === 'string') {
+      return {
+        rel: tag,
+        href: content
+      }
+    } else {
+      let sizes = content.sizes
+      sizes = +sizes || parseInt(sizes)
+      return isNaN(sizes) ? {
+        rel: tag,
+        href: content.href
+      } : {
+        rel: tag,
+        sizes,
+        href: content.href
+      }
+    }
+  } else if (tag === 'apple-touch-startup-image') {
+    return {
+      rel: tag,
+      href: content
+    }
+  } else if (tag === 'apple-mobile-web-app-title') {
+    return {
+      name: tag,
+      content
+    }
+  } else if (tag === 'apple-mobile-web-app-capable') {
+    let value = content
+    if (typeof content === 'boolean' || typeof content === 'number') value = content ? 'yes' : 'no'
+    return {
+      name: tag,
+      content: value
+    }
+  } else if (tag === 'apple-mobile-web-app-status-bar-style') {
+    return {
+      name: tag,
+      content
+    }
+  }
+  return null
+}
+
+export function applyTag (obj, tag, content) {
+  if (!content) return
+  if (obj[tag]) {
+    if (Array.isArray(obj[tag])) {
+      obj[tag].push(content)
+    } else {
+      obj[tag] = [obj[tag], content]
+    }
+  } else {
+    obj[tag] = content
+  }
+}
+
+export function generateHtmlTags (tags) {
+  let html = ''
+  for (let tag in tags) {
+    const attrs = tags[tag]
+    if (Array.isArray(attrs)) {
+      for (let a of attrs) {
+        html = `${html}${generateHtmlTags({
+          [tag]: a
+        })}`
+      }
+    } else {
+      html = `${html}<${tag}`
+      for (let attr in attrs) {
+        html = `${html} ${attr}="${attrs[attr]}"`
+      }
+      html = voidTags.indexOf(tag) === -1 ? `${html}></${tag}>` : `${html} />`
+    }
+  }
+  return html
 }